services/write-service/requirements.txt
aiohttp==3.9.1
aiosqlite==0.19.0
annotated-types==0.7.0
anyio==4.9.0
asyncpg==0.29.0
click==8.1.8
exceptiongroup==1.3.0
fastapi==0.116.1
h11==0.16.0
idna==3.10
psycopg2-binary==2.9.9
pydantic==2.11.7
pydantic_core==2.33.2
pydantic-settings==2.0.3
PyJWT==2.8.0
redis==5.0.1
requests==2.31.0
sniffio==1.3.1
SQLAlchemy==2.0.41
starlette==0.47.1
typing-inspection==0.4.1
typing_extensions==4.14.1
uvicorn==0.35.0
websockets==12.0
bleach==6.2.0
aiofiles==24.1.0
pandas==2.2.2
numpy==1.26.4
Pillow==10.4.0
